Start treating it as a gateway
