+++ web/apps/photos/src/services/entityService.ts
@@ -1,5 +1,4 @@
-import { sharedCryptoWorker } from "@/base/crypto";
-import { decryptMetadataJSON } from "@/base/crypto/ente";
+import { decryptMetadataJSON, sharedCryptoWorker } from "@/base/crypto";
 import log from "@/base/log";
 import { apiURL } from "@/base/origins";
 import HTTPService from "@ente/shared/network/HTTPService";
