+++ web/packages/base/session-store.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { decryptBox } from "./crypto/ente";
+import { decryptBox } from "./crypto";
 import { toB64 } from "./crypto/libsodium";
 
 /**
